requirements.txt
Django
djangorestframework
channels
channels_redis
psycopg2-binary
celery
django-cors-headers
python-decouple
